# solution/openapi.yaml
openapi: 3.0.3
info:
  title: PROD Backend 2025 Advertising Platform API
  version: "1.0.0"
  description: >
    API для управления данными клиентов, рекламодателей, рекламными кампаниями,
    показом объявлений, статистикой и управлением "текущим днём" в системе.
...
paths:
...
  /advertisers/{advertiserId}:
    get:
      tags:
        - Advertisers
      summary: Получение рекламодателя по ID
      description: Возвращает информацию о рекламодателе по его ID.
      operationId: getAdvertiserById
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о рекламодателе успешно получена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Advertiser'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Некорректный идентификатор.
...
  /advertisers/{advertiserId}/campaigns:
    post:
      tags:
        - Campaigns
      summary: Создание рекламной кампании
      description: Создаёт новую рекламную кампанию для указанного рекламодателя.
      operationId: createCampaign
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, для которого создаётся кампания.
          schema:
            type: string
            format: uuid
      requestBody:
        description: Объект с данными для создания рекламной кампании.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignCreate'
      responses:
        '201':
          description: Рекламная кампания успешно создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
    get:
      tags:
        - Campaigns
      summary: Получение рекламных кампаний рекламодателя c пагинацией
      description: Возвращает список рекламных кампаний для указанного рекламодателя с пагинацией.
      operationId: listCampaigns
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, для которого запрашиваются кампании.
          schema:
            type: string
            format: uuid
        - in: query
          name: size
          schema:
            type: integer
          description: Количество элементов на странице.
        - in: query
          name: page
          schema:
            type: integer
          description: Номер страницы.
      responses:
        '200':
          description: Список рекламных кампаний рекламодателя.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
  /advertisers/{advertiserId}/campaigns/{campaignId}:
    get:
      tags: [ Campaigns ]
      summary: Получение кампании по ID
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, которому принадлежит кампания.
          schema:
            type: string
            format: uuid
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, которую необходимо получить.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Кампания успешно получена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Рекламодатель или кампания не найдены.
        '400':
          description: Переданы недействительные данные.
    put:
      tags:
        - Campaigns
      summary: Обновление рекламной кампании
      description: Обновляет разрешённые параметры рекламной кампании до её старта.
      operationId: updateCampaign
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, которому принадлежит кампания.
          schema:
            type: string
            format: uuid
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, которую необходимо обновить.
          schema:
            type: string
            format: uuid
      requestBody:
        description: Объект с обновлёнными данными рекламной кампании.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignUpdate'
      responses:
        '200':
          description: Рекламная кампания успешно обновлена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Рекламодатель или кампания не найдены.
        '400':
          description: Переданы недействительные данные.
    delete:
      tags:
        - Campaigns
      summary: Удаление рекламной кампании
      description: Удаляет рекламную кампанию рекламодателя по заданному campaignId.
      operationId: deleteCampaign
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, которому принадлежит кампания.
          schema:
            type: string
            format: uuid
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, которую необходимо удалить.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Рекламная кампания успешно удалена.
        '404':
          description: Рекламодатель или кампания не найдены.
        '400':
          description: Переданы недействительные данные.
  # Изображения
...
  /stats/campaigns/{campaignId}:
    get:
      tags:
        - Statistics
      summary: Получение статистики по рекламной кампании
      description: Возвращает агрегированную статистику (показы, переходы, затраты и конверсию) для заданной рекламной кампании.
      operationId: getCampaignStats
      parameters:
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, для которой запрашивается статистика.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статистика по рекламной кампании успешно получена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '404':
          description: Кампания не найдена.
        '400':
          description: Переданы недействительные данные.
  /stats/advertisers/{advertiserId}/campaigns:
    get:
      tags:
        - Statistics
      summary: Получение агрегированной статистики по всем кампаниям рекламодателя
      description: Возвращает сводную статистику по всем рекламным кампаниям, принадлежащим заданному рекламодателю.
      operationId: getAdvertiserCampaignsStats
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, для которого запрашивается статистика.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Агрегированная статистика по всем кампаниям рекламодателя успешно получена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
  /stats/campaigns/{campaignId}/daily:
    get:
      tags:
        - Statistics
      summary: Получение ежедневной статистики по рекламной кампании
      description: Возвращает массив ежедневной статистики для указанной рекламной кампании.
      operationId: getCampaignDailyStats
      parameters:
        - in: path
          name: campaignId
          required: true
          description: UUID рекламной кампании, для которой запрашивается ежедневная статистика.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ежедневная статистика по рекламной кампании успешно получена.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyStats'
        '404':
          description: Кампания не найдена.
        '400':
          description: Переданы недействительные данные.
  /stats/advertisers/{advertiserId}/campaigns/daily:
    get:
      tags:
        - Statistics
      summary: Получение ежедневной агрегированной статистики по всем кампаниям рекламодателя
      description: Возвращает массив ежедневной сводной статистики по всем рекламным кампаниям заданного рекламодателя.
      operationId: getAdvertiserDailyStats
      parameters:
        - in: path
          name: advertiserId
          required: true
          description: UUID рекламодателя, для которого запрашивается ежедневная статистика по кампаниям.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ежедневная агрегированная статистика успешно получена.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyStats'
        '404':
          description: Рекламодатель не найден.
        '400':
          description: Переданы недействительные данные.
  # Управление временем
...
  /moderation/requests:
    get:
      tags:
        - Moderation
      summary: Получение запросов на модерацию c пагинацией
      description: Возвращает список запросов на модерацию с пагинацией.
      operationId: listModerationRequests
      parameters:
        - in: query
          name: size
          schema:
            type: integer
          description: Количество элементов на странице.
        - in: query
          name: page
          schema:
            type: integer
          description: Номер страницы.
      responses:
        '200':
          description: Список запросов на модерацию.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModerationRequest'
        '400':
          description: Переданы недействительные данные.
...
  /moderation/reject:
    post:
      tags:
        - Moderation
      summary: Отклонение кампании
      operationId: rejectCampaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
        required: true
      responses:
        '200':
          description: Кампания успешно отклонена.
          content:
            application/json:
              schema:
                type: string
  /moderation/approve:
    post:
      tags:
        - Moderation
      summary: Одобрение кампании
      operationId: approveCampaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
        required: true
      responses:
        '200':
          description: Кампания успешно одобрена.
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
...
    ModerationRequest:
      type: object
      properties:
        ad_title:
          type: string
        ad_text:
          type: string
        campaign_id:
          type: string
          format: uuid
      required:
        - ad_title
        - ad_text
        - campaign_id
    DecisionRequest:
      type: object
      properties:
        campaign_id:
          type: string
          format: uuid
      required:
        - campaign_id
    # --- Клиенты ---
...
    Advertiser:
      type: object
      description: Объект, представляющий рекламодателя.
      properties:
        advertiser_id:
          type: string
          format: uuid
          description: Уникальный идентификатор рекламодателя (UUID).
        name:
          type: string
          description: Название рекламодателя.
      required:
        - advertiser_id
        - name
    # --- ML скор ---
...
    Campaign:
      type: object
      description: Объект, представляющий рекламную кампанию.
      properties:
        campaign_id:
          type: string
          format: uuid
          description: Уникальный идентификатор рекламной кампании (UUID).
        advertiser_id:
          type: string
          format: uuid
          description: UUID рекламодателя, которому принадлежит кампания.
        impressions_limit:
          type: integer
          description: Лимит показов рекламного объявления (фиксируется до старта кампании).
        clicks_limit:
          type: integer
          description: Лимит переходов (кликов) по рекламному объявлению (фиксируется до старта кампании).
        cost_per_impression:
          type: number
          format: float
          description: Стоимость одного показа рекламного объявления.
        cost_per_click:
          type: number
          format: float
          description: Стоимость одного перехода (клика) по рекламному объявлению.
        ad_title:
          type: string
          description: Название рекламного объявления.
        ad_text:
          type: string
          description: Текст рекламного объявления.
        start_date:
          type: integer
          format: int32
          description: День старта показа рекламного объявления (включительно).
        end_date:
          type: integer
          format: int32
          description: День окончания показа рекламного объявления (включительно).
        image_url:
          type: string
          description: Ссылка на изображение
        targeting:
          $ref: '#/components/schemas/Targeting'
      required:
        - campaign_id
        - advertiser_id
        - impressions_limit
        - clicks_limit
        - cost_per_impression
        - cost_per_click
        - ad_title
        - ad_text
        - start_date
        - end_date
        - targeting
    CampaignCreate:
      type: object
      description: Объект для создания новой рекламной кампании.
      properties:
        impressions_limit:
          type: integer
          description: Задаёт лимит показов для рекламного объявления.
        clicks_limit:
          type: integer
          description: Задаёт лимит переходов для рекламного объявления.
        cost_per_impression:
          type: number
          format: float
          description: Стоимость одного показа объявления.
        cost_per_click:
          type: number
          format: float
          description: Стоимость одного перехода (клика) по объявлению.
        ad_title:
          type: string
          description: Название рекламного объявления.
        ad_text:
          type: string
          description: Текст рекламного объявления.
        start_date:
          type: integer
          format: int32
          description: День начала показа рекламного объявления (включительно).
        end_date:
          type: integer
          format: int32
          description: День окончания показа рекламного объявления (включительно).
        targeting:
          $ref: '#/components/schemas/Targeting'
      required:
        - impressions_limit
        - clicks_limit
        - cost_per_impression
        - cost_per_click
        - ad_title
        - ad_text
        - start_date
        - end_date
    CampaignUpdate:
      type: object
      description: Объект для обновления параметров кампании, которые разрешено изменять до старта кампании.
      properties:
        impressions_limit:
          type: integer
          description: Задаёт лимит показов для рекламного объявления.
        clicks_limit:
          type: integer
          description: Задаёт лимит переходов для рекламного объявления.
        cost_per_impression:
          type: number
          format: float
          description: Новая стоимость одного показа объявления.
        cost_per_click:
          type: number
          format: float
          description: Новая стоимость одного перехода (клика) по объявлению.
        ad_title:
          type: string
          description: Новое название рекламного объявления.
        ad_text:
          type: string
          description: Новый текст рекламного объявления.
        start_date:
          type: integer
          format: int32
          description: День начала показа рекламного объявления (включительно).
        end_date:
          type: integer
          format: int32
          description: День окончания показа рекламного объявления (включительно).
        targeting:
          $ref: '#/components/schemas/Targeting'
          description: Новые параметры таргетирования для рекламной кампании.
    Targeting:
      type: object
      description: Объект, описывающий настройки таргетирования для рекламной кампании.
      properties:
        gender:
          type: string
          enum: [ MALE, FEMALE, ALL ]
          nullable: true
          description: Пол аудитории для показа объявления (MALE, FEMALE или ALL).
        age_from:
          type: integer
          nullable: true
          description: Минимальный возраст аудитории (включительно) для показа объявления.
        age_to:
          type: integer
          nullable: true
          description: Максимальный возраст аудитории (включительно) для показа объявления.
        location:
          type: string
          nullable: true
          description: Локация аудитории, для которой будет показано объявление.
    # --- Рекламное объявление ---
...
    Stats:
      type: object
      description: Объект, содержащий агрегированную статистику для рекламной кампании или рекламодателя.
      properties:
        impressions_count:
          type: integer
          description: Общее количество уникальных показов рекламного объявления.
        clicks_count:
          type: integer
          description: Общее количество уникальных переходов (кликов) по рекламному объявлению.
        conversion:
          type: number
          format: float
          description: Коэффициент конверсии, вычисляемый как (clicks_count / impressions_count * 100) в процентах.
        spent_impressions:
          type: number
          format: float
          description: Сумма денег, потраченная на показы рекламного объявления.
        spent_clicks:
          type: number
          format: float
          description: Сумма денег, потраченная на переходы (клики) по рекламному объявлению.
        spent_total:
          type: number
          format: float
          description: Общая сумма денег, потраченная на кампанию (показы и клики).
      required:
        - impressions_count
        - clicks_count
        - conversion
        - spent_impressions
        - spent_clicks
        - spent_total
    DailyStats:
      allOf:
        - $ref: '#/components/schemas/Stats'
        - type: object
          description: Объект, представляющий ежедневную статистику с указанием дня.
          properties:
            date:
              type: integer
              format: int32
              description: День, за который была собрана статистика.
          required:
            - date
...